Displaying filters after fetching data
