+++ src/components/PeoplePage.tsx
@@ -1,7 +1,30 @@
+import { useEffect, useState } from 'react';
 import { PeopleFilters } from './PeopleFilters';
 import { PeopleTable } from './PeopleTable';
+import { getPeople } from '../api';
+import { Person } from '../types';
 
 export const PeoplePage = () => {
+  const [people, setPeople] = useState<Person[]>([]);
+  const [loading, setLoading] = useState<boolean>(true);
+  const [error, setError] = useState<boolean>(false);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const data = await getPeople();
+
+        setPeople(data);
+      } catch (e) {
+        setError(true);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchData();
+  }, []);
+
   return (
     <>
       <h1 className="title">People Page</h1>
@@ -9,12 +32,14 @@ export const PeoplePage = () => {
       <div className="block">
         <div className="columns is-desktop is-flex-direction-row-reverse">
           <div className="column is-7-tablet is-narrow-desktop">
-            <PeopleFilters />
+            {!loading && <PeopleFilters />}
           </div>
 
           <div className="column">
             <div className="box table-container">
-              <PeopleTable />
+              <p>There are no people matching the current search criteria</p>
+
+              <PeopleTable people={people} loading={loading} error={error} />
             </div>
           </div>
         </div>
